bb-imager-gui: Start separating UI
- Make code easier to modify by separating UI and business logic
- Do not directly use fields in state from the UI.
