+++ bb-imager-gui/src/main.rs
@@ -4,10 +4,10 @@ use std::{collections::HashSet, time::Duration};
 
 use bb_config::config;
 use constants::PACKAGE_QUALIFIER;
-use helpers::ProgressBarState;
-use iced::{Element, Subscription, Task, futures::SinkExt, widget};
+use helpers::{FlashingCustomization, ProgressBarState};
+use iced::{Subscription, Task, futures::SinkExt, widget};
 use message::BBImagerMessage;
-use pages::{Screen, configuration::FlashingCustomization};
+use pages::Screen;
 use tokio_stream::StreamExt;
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
@@ -16,6 +16,7 @@ mod constants;
 mod helpers;
 mod message;
 mod pages;
+mod ui;
 
 fn main() -> iced::Result {
     tracing_subscriber::registry()
@@ -51,7 +52,7 @@ fn main() -> iced::Result {
         ..Default::default()
     };
 
-    iced::application(constants::APP_NAME, BBImager::update, BBImager::view)
+    iced::application(constants::APP_NAME, BBImager::update, ui::view)
         .subscription(BBImager::subscription)
         .theme(BBImager::theme)
         .window(settings)
@@ -70,17 +71,13 @@ struct BBImager {
     selected_board: Option<usize>,
     selected_image: Option<helpers::BoardImage>,
     selected_dst: Option<helpers::Destination>,
-    destinations: HashSet<helpers::Destination>,
-    search_bar: String,
+    destinations: Vec<helpers::Destination>,
     cancel_flashing: Option<iced::task::Handle>,
     customization: Option<FlashingCustomization>,
-    flashing_state: Option<pages::flash::FlashingState>,
+    flashing_state: Option<pages::FlashingState>,
 
     timezones: widget::combo_box::State<String>,
     keymaps: widget::combo_box::State<String>,
-
-    // Flag to indicate if destinations are selectable
-    destination_selectable: bool,
 }
 
 impl BBImager {
@@ -137,13 +134,11 @@ impl BBImager {
                     .map(|x| x.to_string())
                     .collect(),
             ),
-            destination_selectable: true,
             screen: Vec::with_capacity(3),
             selected_board: Default::default(),
             selected_image: Default::default(),
             selected_dst: Default::default(),
             destinations: Default::default(),
-            search_bar: Default::default(),
             cancel_flashing: Default::default(),
             customization: Default::default(),
             flashing_state: Default::default(),
@@ -238,8 +233,6 @@ impl BBImager {
                 self.selected_image = Some(x);
                 self.screen.clear();
                 self.screen.push(Screen::Home);
-
-                return self.refresh_destinations();
             }
             BBImagerMessage::SelectLocalImage(flasher) => {
                 let extensions = helpers::file_filter(flasher);
@@ -267,7 +260,6 @@ impl BBImager {
                 self.selected_dst.take();
                 self.selected_image.take();
                 self.selected_board.take();
-                self.search_bar.clear();
                 self.destinations.clear();
             }
             BBImagerMessage::SwitchScreen(x) => {
@@ -286,16 +278,13 @@ impl BBImager {
                 tracing::debug!("Pop screen");
                 self.screen.pop();
             }
-            BBImagerMessage::Search(x) => {
-                self.search_bar = x;
-            }
             BBImagerMessage::CancelFlashing => {
                 if let Some(task) = self.cancel_flashing.take() {
                     task.abort();
                 }
 
                 if let Some(x) = &self.flashing_state {
-                    if let Some(y) = x.progress.cancel() {
+                    if let Some(y) = x.progress().cancel() {
                         return Task::done(BBImagerMessage::StopFlashing(y));
                     }
                 }
@@ -329,10 +318,9 @@ impl BBImager {
                 return Task::batch([progress_task, notification_task]);
             }
             BBImagerMessage::Destinations(x) => {
-                if !self.destination_selectable {
+                if !self.is_destionation_selectable() {
                     assert_eq!(x.len(), 1);
-                    let temp: Vec<&helpers::Destination> = x.iter().collect();
-                    self.selected_dst = Some(temp[0].clone());
+                    self.selected_dst = Some(x[0].clone());
                 }
                 self.destinations = x;
             }
@@ -377,7 +365,6 @@ impl BBImager {
     }
 
     fn push_page(&mut self, x: Screen) -> Task<BBImagerMessage> {
-        self.search_bar.clear();
         self.screen.push(x.clone());
 
         match x {
@@ -387,7 +374,7 @@ impl BBImager {
             Screen::ImageSelection(page) => {
                 tracing::info!("Image Selection Screen");
                 let board = self.selected_board.unwrap();
-                return self.fetch_remote_subitems(board, &page.idx);
+                return self.fetch_remote_subitems(board, page.idx());
             }
             _ => {}
         }
@@ -462,126 +449,8 @@ impl BBImager {
         Task::batch(remote_image_jobs)
     }
 
-    fn view(&self) -> Element<BBImagerMessage> {
-        tracing::debug!("Page Stack: {:#?}", self.screen);
-
-        match self.screen.last().expect("No Screen") {
-            Screen::Home => pages::home::view(
-                self.selected_board.map(|x| self.boards.device(x)),
-                self.selected_image.as_ref(),
-                self.selected_dst.as_ref(),
-                self.destination_selectable,
-            ),
-            Screen::BoardSelection => {
-                pages::board_selection::view(&self.boards, &self.search_bar, &self.downloader)
-            }
-            Screen::ImageSelection(page) if page.flasher == config::Flasher::SdCard => {
-                let board = self.selected_board.expect("Missing Board");
-                let flasher = page.flasher;
-
-                page.view(
-                    self.boards.images(board, &page.idx),
-                    &self.search_bar,
-                    &self.downloader,
-                    [
-                        pages::image_selection::ExtraImageEntry::new(
-                            "Custom Image",
-                            constants::FILE_ADD_ICON,
-                            BBImagerMessage::SelectLocalImage(flasher),
-                        ),
-                        pages::image_selection::ExtraImageEntry::new(
-                            "Format Sd Card",
-                            constants::FORMAT_ICON,
-                            BBImagerMessage::SelectImage(helpers::BoardImage::SdFormat),
-                        ),
-                    ]
-                    .into_iter(),
-                )
-            }
-            Screen::ImageSelection(page) => {
-                let board = self.selected_board.expect("Missing Board");
-                let flasher = page.flasher;
-
-                page.view(
-                    self.boards.images(board, &page.idx),
-                    &self.search_bar,
-                    &self.downloader,
-                    [pages::image_selection::ExtraImageEntry::new(
-                        "Custom Image",
-                        constants::FILE_ADD_ICON,
-                        BBImagerMessage::SelectLocalImage(flasher),
-                    )]
-                    .into_iter(),
-                )
-            }
-            Screen::DestinationSelection => {
-                pages::destination_selection::view(self.destinations.iter(), &self.search_bar)
-            }
-            Screen::ExtraConfiguration => pages::configuration::view(
-                self.customization.as_ref().unwrap(),
-                &self.timezones,
-                &self.keymaps,
-            ),
-            Screen::Flashing => pages::flash::view(
-                self.flashing_state.as_ref().unwrap(),
-                self.cancel_flashing.is_some(),
-            ),
-            Screen::FlashingConfirmation => {
-                let base = pages::home::view(
-                    self.selected_board.map(|x| self.boards.device(x)),
-                    self.selected_image.as_ref(),
-                    self.selected_dst.as_ref(),
-                    self.destination_selectable,
-                );
-                let menu = widget::column![
-                    widget::text("Would you like to apply customization settings?"),
-                    widget::row![
-                        widget::button("Edit Settings")
-                            .on_press(BBImagerMessage::ReplaceScreen(Screen::ExtraConfiguration)),
-                        widget::button("Yes").on_press(BBImagerMessage::StartFlashing),
-                        widget::button("No")
-                            .on_press(BBImagerMessage::StartFlashingWithoutConfiguraton),
-                        widget::button("Abort")
-                            .on_press(BBImagerMessage::SwitchScreen(Screen::Home))
-                    ]
-                    .spacing(8)
-                ]
-                .align_x(iced::Alignment::Center)
-                .padding(16)
-                .spacing(16);
-
-                let menu = widget::container(menu)
-                    .style(|_| widget::container::background(iced::Color::WHITE));
-
-                let overlay = widget::opaque(widget::center(menu).style(|_| {
-                    widget::container::background(iced::Color {
-                        a: 0.8,
-                        ..iced::Color::BLACK
-                    })
-                }));
-                widget::stack![base, overlay].into()
-            }
-        }
-    }
-
     const fn theme(&self) -> iced::Theme {
         iced::Theme::Light
-    }
-
-    fn refresh_destinations(&mut self) -> Task<BBImagerMessage> {
-        if let Some(flasher) = self.flasher() {
-            self.destination_selectable = helpers::is_destination_selectable(flasher);
-
-            // Do not use subscription for static destinations
-            if !self.destination_selectable {
-                return Task::perform(
-                    async move { helpers::destinations(flasher).await },
-                    BBImagerMessage::Destinations,
-                );
-            }
-        }
-
-        Task::none()
     }
 
     fn config(&self) -> FlashingCustomization {
@@ -601,7 +470,7 @@ impl BBImager {
             .boards
             .device(self.selected_board.expect("Missing board"))
             .documentation;
-        self.flashing_state = Some(pages::flash::FlashingState::new(
+        self.flashing_state = Some(pages::FlashingState::new(
             ProgressBarState::PREPARING,
             docs_url.as_ref().map(|x| x.to_string()).unwrap_or_default(),
         ));
@@ -652,11 +521,17 @@ impl BBImager {
         if let Some(flasher) = self.flasher() {
             // Do not use subscription for static destinations
             // Also do not use subscription when on other screens. Can cause Udisk2 to crash.
-            if self.destination_selectable
-                && self.screen.last().expect("No screen") == &Screen::DestinationSelection
+            if self.is_destionation_selectable()
+                && self
+                    .screen
+                    .last()
+                    .expect("No screen")
+                    .is_destination_selection()
             {
                 let stream = iced::futures::stream::unfold(flasher, move |f| async move {
-                    let dsts = helpers::destinations(flasher).await;
+                    let mut dsts: Vec<helpers::Destination> =
+                        helpers::destinations(flasher).await.into_iter().collect();
+                    dsts.sort_by_key(|a| a.size());
                     let dsts = BBImagerMessage::Destinations(dsts);
                     Some((dsts, f))
                 })
@@ -675,5 +550,61 @@ impl BBImager {
         }
         let dev = self.boards.device(self.selected_board?);
         Some(dev.flasher)
+    }
+
+    pub(crate) fn selected_device(&self) -> Option<&config::Device> {
+        self.selected_board.map(|x| self.boards.device(x))
+    }
+
+    pub(crate) fn selected_image(&self) -> Option<&helpers::BoardImage> {
+        self.selected_image.as_ref()
+    }
+
+    pub(crate) fn selected_destination(&self) -> Option<&helpers::Destination> {
+        self.selected_dst.as_ref()
+    }
+
+    pub(crate) fn is_destionation_selectable(&self) -> bool {
+        if let Some(flasher) = self.flasher() {
+            helpers::is_destination_selectable(flasher)
+        } else {
+            false
+        }
+    }
+
+    pub(crate) fn devices(&self) -> impl Iterator<Item = (usize, &config::Device)> {
+        self.boards.devices()
+    }
+
+    pub(crate) fn images(&self, idx: &[usize]) -> Option<Vec<(usize, &config::OsListItem)>> {
+        self.boards.images(self.selected_board.unwrap(), idx)
+    }
+
+    pub(crate) fn destinations(&self) -> impl Iterator<Item = &helpers::Destination> {
+        self.destinations.iter()
+    }
+
+    pub(crate) fn downloader(&self) -> &bb_downloader::Downloader {
+        &self.downloader
+    }
+
+    pub(crate) fn customization(&self) -> Option<&FlashingCustomization> {
+        self.customization.as_ref()
+    }
+
+    pub(crate) fn timezones(&self) -> &widget::combo_box::State<String> {
+        &self.timezones
+    }
+
+    pub(crate) fn keymaps(&self) -> &widget::combo_box::State<String> {
+        &self.keymaps
+    }
+
+    pub(crate) fn flashing_state(&self) -> Option<&pages::FlashingState> {
+        self.flashing_state.as_ref()
+    }
+
+    pub(crate) fn is_flashing(&self) -> bool {
+        self.cancel_flashing.is_some()
     }
 }
